fix: Solve false positive attention markers
Under some conditions (see savefile in `scratch` file),
melee units were reporting that they could attack next turn
but then wouldn't.  This is because there are two places where
the attack is calculated, once in planningView which adds
the attentionMarker if it predicts an attack; 2. in
Underworld.executeNPCTurn.  However since I'm now caching
the targets and if they are able to attack, it was running
this.canAttackTarget on a unit that hadn't had their stamina refreshed
 yet. and they shouldn't have their stamina refreshed yet because the
 ranged units have to move before they melee units should move.

Will refactor this to more sustainable solution in next commit.