Revert "ref: Guard proper unit movement"
This method caused unforseen complexity.  see explanation in
15dc138f.  Better to not have to sync a whole new part of underworld
which introduces room for error and I saw both with the archer
sync resetting and the players not being able to move
on their turn if the activeUnitSubType wasn't synced.
TurnPhase syncing has been working well for a long time now,
I'll just stick with that and manage stamina for allowing units to
move on their turn.

Also caching the target should still prevent false positive and
false negative unit predictions.

This reverts commit fad5062e07340722bec417dc5973b9231780328d.