ref: Archer movement so it can move to two points
while only awaiting movement once.
If a unit.action() function awaits moveTowards more than once
it will work on singleplayer but cause a racetimeout on multiplayer
on the server.
This is because the server's gameloop executes immediately and fully
until no units are left processing and it detects this by seeing
if a unit no longer has a path or if they are out of stamina,
and so when I was using two `await moveTowards()` when the
first one finished the path would be empty and the headless
server would assume the unit was done processing and stop
running the gameloop.

So when you need to path a unit to two specific points in one turn
(like the archer getting LOS then getting closer in range) use
Unit.moveTowardsMulti