optimize: Remove extraneous call to moveMouse
The functionality here got coupled over time resulting in
mouseMove / runPredictions being invoked both every time the mouse was moved and
every game loop.
runPredictions can be quite expensive when invoked with a long running
prediction and due to 78421412fd9eb8406f6fd947f9c1d483166d279a
I was calling it every gameloop but the reasoning from that
commit no longer applies so it is not needed in gameLoop to show
the line from the player cast to the target which no longer exists.

So i've split up the responsibilities of the mouseMove function
and runPredictions.
Now runPredictions is only callled when:
- A spell changes (CardUI.ts)
- The current player moves (eventListeners.ts)
- Another player moves (networkHandler.ts)
- The mouse moves (views.ts)

and the mouseMove function has been renamed to useMousePosition
since it needs to be invoked both when the mouse moves
and in the gameLoop (so the player unit can move)

The big effect of this commit is that runPredictions will no longer run
every loop and only run when needed