fix: (important) stacking spells expense scaling
When I added mana refunds I moved the `cardUsage +=`
out of it's own forloop and into the existing forloop
within the refund logic, however, I didn't notice that
I moved it below where quantity logic prevents
stackable cards from each executing the loop and so
it was only increasing cardUsage once for a stack of
cards that supports quantity.

This fix ensures that card expence will stay at what it's
supposed to be after cast for stackable cards.