ref: Use casterPositionAtTimeOfCast instead of
"You cannot move while casting".
Preventing movement while casting was added friction in an attempt to prevent the
prediction error of casting a burst or arrow spell and then moving before
the spell triggered which would result in unexpected result.
However, this friction got in the way at all times just to prevent an
issue with 2  types of spells - so instead, I've added the
effecState.casterPositionAtTimeOfCast property which can be used for spells
that depend on the caster position.  This is more dev effort to maintain, but
a much better user experience