ref: Triggering pickups
Separate Pickup.effect() into effect and willTrigger so I can determine
in a prediction way if the pickup will trigger for a given unit without
actually triggering anything.  This is critical for the current and new way
of always triggering pickups via the network because in the previous commits
it made it so the pickup was triggering even in prediction because there
was no existing way to actually test them with prediction since many use
a reference to `player` and there is no prediction version of player.
This allows them to actually "predict" if they will trigger.
Split triggerPickup into tryTriggerPickup which causes the network
message if it would trigger and triggerPickup which actually triggers
it.