fix: infinite stalled loop
Reproducable before this commit, by killing all enemies,
spawning 1 ally, then quitting a multiplayer game.

It would loop indefinitely because tho stalled loop was broadcast (and
would be set asyncronously) there was an NPC_ALLY message in the queue that would
trigger next causing an infinite loop.  This change sets STalled immediately instead
of broadcasting (theres no one to broadcast to anyway if its stalled.
My last idea if this didn't work was to clear the queue before setting it to stalled, but
this worked.