ref: Intentional seeded randomization
ref: Refactor random number generation functions to take seeded
random state optionally, to allow for use in a non-seeded manner
for invokations that don't need to be seeded, like visual effects

fix: Remove uses of seed where it does not belong.  The seed is intended to
replicate the same random number across all clients, but if I use
it for things that don't need to be seeded (such as particle effects),
the changes of the random state getting out of sync is drastic.