ref: Take Pickup.create off the network.
Now that seed syncing is resolved, pickups should be
consistant and moving them to the network
was causing problems of their own.
For example with Deathmasons red portals
being overwritten because it would sent a create message that waited in
queue while SET_PHASE took a snapshot of
the state and that state would overwrite the newly
created portals.

Note: AQUIRE_PICKUP is still networked