fix: Pickup; protect against triggering
ENTER_PORTAL multiple times.  Prevent sending multiple AQUIRE_PICKUP messages.

This commit along with 372072cd fixes the Double Spend issue with pickups.
Even after 372072cd, non single-use pickups such as Portals were still triggering
multiple times.  This change removes
singleUse to ensure that pickups are only
ever triggered once, and it also prevents
bogging down the network with redundant
AQUIRE_PICKUP messages because flaggedForRemoval
wont be set to true until the client processes the AQUIRE_PICKUP,
so it would keep sending them (all but the first would short circuit)
until the first processes.  This is a waste and so I just have it
triggerPickup locally as well as sending the AQUIRE_PICKUP message
which will make sure that it triggers on all clients even if one of the clients
didn't detect the collision.
And the fact that triggerPickup won't trigger for a pickup that has been
flaggedForRemoval (and now that there aren't any non singleUse pickups),
the AQUIRE_PICKUP message won't do anything if it has already processed
a triggering for a pickup locally.