fix: Auto backup save from saving
too late if it needed to overwrite.
The user won't see the prompt until after reconnection, and if it's asking permission
to overwrite it would then save the newly generated level
instead of the one that was just disconnected. This change ensures it
saves immediately even if it overwrites a previous backup save so it will
save the state of the game before the server disconnect.