fix: desync issue
where player loses control of themself.
This was reproducable by having 2 players in a multiplayer came, killing one unit and then
casting Harvest on them.  Then the underworld's units array would be
filled with a bunch of units that weren't cleaned up and some were
the player units that still shared the same id, but these 2 changes (changing the
id's of cleaned up units) and removing cleaned up units right
after a sync resolves this issue.