fix: Issue where spawning
a Gripthulu miniboss in a pool of blood
resulting in him dying but having full health.

This happened because if a unit would die when being
placed in liquid, since unit.create checks for liquid
collisions it would die BEFORE being made miniboss.
Now the unit is made miniboss BEFORE the liquid check