ref: Pickups
Headless server is the source of truth for pickup collisions.
However, this is complex because headless server processes forceMoves
instantly, we don't want pickups to trigger on the client side before
the animations have completed.  Therefore, when headless sends the
QUEUE_PICKUP_TRIGGER message, the clients store the pickup info
in a queue.  And once, on the client, the unit collides with the pickup,
if the pickup information is in the queue, it THEN triggers the pickup.