ref: Pickup
Handle edgecase where player touches pickup on client before server has.
Usually server processes the touched pickup first because it triggers
all it's movement loops immediately; however, it's possible due to a desync
or maybe due to large latency while a player is moving that the player
passess through the pickup on the client before it does on the server.
In that case, if the server adds to the aquirePickupQueue after the
client unit has already passed through then it won't trigger on the client
(it will timeout), so this edge case has the client trigger the pickup for
everyone. (so there's no desync).