ref: join game as player
Joining a game with the same
name as a disconnected player who has a different clientId will now
automatically switch that player to the other one.
This drastically improves the experience of joining saved games
where you're trying to assume control of your saved player
but your clientId has changed.