ref: Redo syncPickups
to account for dash to pickup throwing
`Attempted to aquire queued pickup via timeout but pickup is undefined`
Because the pickup was used and removed on server
and then SYNC_SOME_STATE was sent to the client
which synced, removing the pickup while the
pickup's id was still in its aquirePickupQueue array.

And I couldn't just not remove the pickups that were in that
array because the pickup could be overwritten in the next
for loop where the pickup arrays were synced index by index
instead of id by id. So now it syncs id by id