balance: send_mana: Allow sending mana to non mana users
Per SoulMuncher:
One interesting thing you could do with the spell previously was create a "mana battery" with a decoy or other cheap summon. Give it your extra mana this turn and steal it back next turn or in a few. You could also use it to make golems and other melee units vulnerable to spells like mana burn. I definitely favor the idea of letting the player send mana to anything, especially when thinking about all of the other combos and strategies that might develop as more mana-based spells are added.