chore: Add spellmasons-mods files to repo
This only adds the folders that are used, not the entire spellmasons-mods repo.
This is so that new users of this repo don't have to copy the spellmasons-mods repo
for this repo to build. Also it will ensure that changes to the spellmasons-mods repo
are committed to this repository so that anyone can `npm build` from this repo
without having to sync the spellmasons-mods repo