src: Hotseat player handling
Start with player 1 on each new level.
Wait to switch to player 1's turn until after the NPCs have taken their turn.

Move changeToNextHotseatPlayer logic to
AFTER tryGameOver and tryEndPlayerTurnPhase
to ensure that either of those cases are checked for before changing to next player.
This is because if a player is changed to but they are unable to act, it will end their turn,
which invokes endPlayerTurn and the
nextPhase and gameOver checks need to occur before trying to switch AGAIN to another
hotseat player