ref: Clients are the source of truth for their player's state
for during game syncs.
The server is the source of truth when LOADing a game or when changing which player the client is controlling.

This should resolve the issue of player's losing Summon cards
or weird behavior like having your upgrade options regenerate if your ally level's up while your choosing your upgrades.

Resolves #188