fix: FORCE_TRIGGER_PICKUP
In a tragic misunderstanding on my part, as some point I changed
triggerPickup to get passed fromPlayer; however
the FORCE_TRIGGER_PICKUP message can be sent from ANY CLIENT for any unit.
Which means that any time a FORCE_TRIGGER_PICKUP was sent, if it first arrived from the client who's player's unit did not touch the pickup, it would
switch which player unit triggered the pickup.

Note: It is built in that duplicate force pickup messages for pickups will be discarded after the first triggers.

Clients may now send FORCE_TRIGGER_PICKUP for their own player unit or a non player unit but not each other's player units because
a client is it's own players' unit's source of truth
and we don't want client B to trigger a pickup for client A just because client B's copy
of the player A's unit is slightly desynced.

Resolves #185