src: Rework arrowEffect
so that it generates predictions on a !prediction
invokation so that all clients will
have a predictedArrowCollisions array (not just the
client who cast the spell)
and there is no higher-scrope predictedArrowCollisions array
stored that could leak into other invokatiosn of runArrowEffect.

TODO: @SoulMuncher, you may not need the `currentState` nor `nextArrowIndex`
now that predictedArrowCollisions is isolated to each real invokation.