fix: Some forceMoves not being properly awaited
which caused desyncs.
This was the cause of many force move desyncs.  It was possible for there to be
a forceMove instance added to underworld.forceMove and then
awaitForceMoves would be called BEFORE the forceMovePromise was refreshed
which means awaitForceMoves returned immediately before the force moves were
actually resolved.

Now, any time a forceMove is added, if the promise doesn't already exist it is refreshed.

Fixes: #244