fix: Let slash animation finish
even if there is a follow up spell.
The removed line in this commit was
added LONG ago to account for a spell
(protect) which removed a targeting circle.
I think this functionality is long obsolete
but what it was currently doing was removing
the slash animation (and others) before it was
completed.

Now animations that are added via a card effect
will stick around until castCards has completed
or until they remove themselves because they
have completed animating.