fix: END_TURN softlock.
Since clients are the source of truth of their player objects,
if the server had to reset a player for being out of bounds
but the client didn't, then the server would have player.isSpawned = false
and the client would not causing a softlock on end turn.

Also added protect in the end turn handler because any player ending
their turn is logically connected and spawned.