fix: Players suddenly at -1000,-1000
Old code was attempting to let players rechoose their
spawn if they got out of bounds
but this unfortunately just would make them
be not-choosing spawn and at -1000,-1000.

I also tried just setting isSpawned to false
to let them rechoose spawn but that results
in all the enemies dying.
So let's just not do anything if they're out of bounds.

Fixes: #457