test: Reports live in console
if promises are left unresolved by the time
the game moves onto casting the next card.effect.

This is a work in progress and could stand to be cleaned up, but it should be super useful for finding cards
with dangling unawaited promises.

Note, some promises are intentionally unawaited like
heal's floatingText.  The effect of the spell need not await the completion of the floating text, so it is ignored.