fix: Units being removed for being "out of bounds"
when they were just close up against an upper wall.

Unit's center point is the center of their image, and since this game is 2D but
allow's "tile overlap" because it is appearing as isometric, the pathing bounds
are pulled up by 10 (-10) and so when checking if a units' center point is
inside a wall, we need to + 10 (down) to account for the fact that they can overlap
by 10 with a wall that is above them.