ref: Remove cooldown
After considering, I've determined that the idea of a cooldown itself is
against the design of Spellmasons.  This was just a bandaid to
a few select spells being overpowered.  I've found it better to
alter the spell logic itself (in the case of freeze to not allow stacking)
and to make them more mana expensive (in the case of resurrect)
which will prevent them from just being cast turn after turn.

Also the modifier itself can prevent the unit from being
refrozen (see freeze) which is much better than cooldowns
 because it also prevents refreezing in multiplayer.