UX: Improve Rune descriptions
In the Rune Upgrade Book: the description ALWAYS
shows what a SINGLE UPGRADE will change.  For example:
always "+20% health" so that you know what is changing when you
click the upgrade button.  If it said the current value "+80% health"
some people might assume that clicking it will yield +80% and get
confused.

The **tooltip description** however shows the current value:
"Revitalize" (newline)
"+80% health on healing"
for example, so you know how much you currently have.

The rune cost has "sp" next to it.  This could be changed to an icon.
This denotes that it is a cost so it's not just a number.
And on hover it shows how much you'll gain when you upgrade.