src: Rune tooltips
On hover plus button only show the current value if it's not 0.
Tooltips are now automatic (and standardized) for modifiers with `unitOfMeasure` property.

Modifiers must not include a dynamic description.  It should read as
"Thorns (the modifier key) damages attackers on hit (the static modifier description)"
This will make translation a lot easier and serve as a sensible
description for BOTH the tooltip
and the rune book