Rune descriptions (#935)
* UX: Improve Rune descriptions

In the Rune Upgrade Book: the description ALWAYS
shows what a SINGLE UPGRADE will change.  For example:
always "+20% health" so that you know what is changing when you
click the upgrade button.  If it said the current value "+80% health"
some people might assume that clicking it will yield +80% and get
confused.

The **tooltip description** however shows the current value:
"Revitalize" (newline)
"+80% health on healing"
for example, so you know how much you currently have.

The rune cost has "sp" next to it.  This could be changed to an icon.
This denotes that it is a cost so it's not just a number.
And on hover it shows how much you'll gain when you upgrade.

* src: Add unitOfMeasure
so that tooltips can be automatically
updated

* src: Rune tooltips

On hover plus button only show the current value if it's not 0.
Tooltips are now automatic (and standardized) for modifiers with `unitOfMeasure` property.

Modifiers must not include a dynamic description.  It should read as
"Thorns (the modifier key) damages attackers on hit (the static modifier description)"
This will make translation a lot easier and serve as a sensible
description for BOTH the tooltip
and the rune book

* ref: Change modifier description back to string

so it can be static

* fix: undefined unit of measure

* css: Center rune names and descriptions

* fix: names for basic stat runes

* copy: Clarify on hover when

upgrading to max by showing the current value and an arrow to "Max"

* css: Fix alignment for plus button

on runes page
due to the added symbol "sp" denoting the cost

* i18n: Runes

Note: Translation should NOT be done in the rune definition
because it needs to be called at runtime because the language may change while the game is running

* i18n: Add localization ids to runes

* fix: forgot to remove updateTooltip from overheal