ref: Ordered modifier events
Modifier events must be ordered so that
their effects are not dependent on the order in
which the modifier was added to the unit.

Tested in cards/index.test.ts
Validated by adding poison and bloodcurse to a unit in any order
and in the events array poison always comes first now.

TODO: Actually specify the orders for all units.
Closes: #933