optim: Chunk getSmartTargets
so it calculates only a chunk of units at a time.  This ensures that the FPS doesn't tank
due to lots of units calculating their targeting every time you move.

When they need to calculate targets to attack on their turn, however,
it does need to calculate them all.

todo: ensuring the attentionMarkers don't flicker could stand to be improved.