fix: Determine rune cost on client
Fixes "runes being undone" in multiplayer.
I never found the exact cause but it seems like it's blocked due to overspend and only on discounted runes so I suspect that there's a desync between the server cost and the client but I was unable to induce it.

So this commit sends the cleint cost along with the upgrade message so it will only ever block overspend on the client.
I also added contextual logging to investigate more
thoroughly when a desync does occur so hopefully i can find the root cause even though this commit will resolve the bug itself.

Resolves #1125